Extract import_derived_class
